Initial Commit
Commit of python program that uses Tensor Flow to train a model capable of recognizing handwritten digits contained in the mnist dataset. Contains the data file you will need, the environment file you will need, and a "how to" document for using all of this code on the Ptolemy cluster at Mississippi State University.